backend/simple_requirements.txt
flask==2.3.3
openai==1.1.1
requests==2.31.0
python-dotenv==1.0.0
elevenlabs==0.2.24
werkzeug==2.3.7 